refactor: renomear app.js para page.js
