--- projeto_pessoal/app.js
@@ -1,30 +1,0 @@
-const app = document.getElementById("app");
-
-function Header({ title }) {
-    return <h1>{title ? title : 'Default Title 2'}</h1>;
-}
-
-function HomePage() {
-    const names = ['Ada Lovelace', 'Grace Hopper', 'Margaret Hamilton'];
-
-    const [likes, setLikes] = React.useState(0);
-
-    function handleClick() {
-        setLikes(likes + 1)
-    }
-    
-    return (
-        <div>
-            <Header title="Develop. Preview. Ship."/>
-            <ul>
-                {names.map((name) => (
-                    <li key={name}>{name}</li>
-                ))}
-            </ul>
-            <button onClick={handleClick}>Like ({likes})</button>
-        </div>
-    );
-}
-
-const root = ReactDOM.createRoot(app);
-root.render(<HomePage/>);
